Fix compilation on AIX: -I/usr/include cannot be used.
The reason is that the C++ compiler needs to add -I/usr/vacpp/include
before -I/usr/include, so our adding of -I/usr/include changes the order.
This causes a compilation error in the C++ header xlocinfo.h.

In any case, these checks done in qmake are unnecessary. If the OpenSSL
headers are in one of those include paths, they will be found on their
own already.

Reviewed-By: Bradley T. Hughes