QCheckBox on an out-of-process server wont allow <SPACE> key & Focus rect is not drawn correctly.
Depending on the type of server, QAxClientSite::TranslateAccelerator()
needs to process the message differently. For ActiveQt based
in-process-servers, only normal Qt event handling is required. For
ActiveQt based out-of-process-servers, the message has to be forwarded
and Qt event handling needs to continue. For all other type of servers,
forward the message and stop Qt event processing.

Styles use the WA_KeyboardFocusChange attribute set on the window to
decide on drawing the focus rect. ActiveQt handles the VK_TAB key in the
QAxServerBase::TranslateAcceleratorW(), the attibute is now set when
focus is changed.

Task-number: 253763
Reviewed-by: Volker Hilsheimer