Fixed font attribute inheritence, text and textArea elements in QtSvg.
Text used to be formatted during the parsing of the SVG file, but
because the text can be referenced by a 'use' element, the text
formatting is not known at this point in time. Now, font attributes can
be inherited from 'use' elements, and the text is formatted each time
it is drawn.

Reviewed-by: Tor Arne