attempt to return something sensible in uncaughtExceptionBacktrace()
JSC doesn't provide a way to get the backtrace after the exception
has happened (you can only get it at the point it's thrown, by
installing a debugger). As a least effort, we try to use the
uncaught exception to provide a 1-line backtrace (the filename and
linenumber of the innermost call where the exception happened).