Fix compilation: add #include <new> if placement new is used
I'm not sure this is the proper fix, though. xlC 7 complains with:

"/usr/vacpp/include/new", line 97.10: 1540-1298 (I) "void *operator
new(size_t, void *) throw()" needs to be declared in the containing
scope to be found by name lookup.

Also add some calculations to the padding in the data structure.

Reviewed-by: Trust Me