Restored window title after another top-level window was closed.
Task: 249995

The window title in S60 is global entity, i.e. is located in application
status pane. When top level window, for example mesage box is closed,
the window title has to be restored for window which is getting the focus.
This way we don't leave title from destroyed window to title bar.

It is also into question, whether the non-maximized dialogs should set
title to statuspane at all. For now they will, but things may change if
we switch to use native dialogs.