Remove QScriptContextPrivate, QScriptContext is now a JSC::CallFrame
There will not be instances of QScriptContext anymore. Pointer to
QScriptContext are just pointer to JSC::CallFrame

This simplifies the code as we do not need to manage the memory for the
QScriptContext anymore.

It is compatible because it is not possible to create
QScriptContext. They were all created by QScriptEngine.
QScriptContext constructor is private.

Aknoweldged-by: Kent