Change the pointer-tracking code to work everywhere.
Currently, if you create a QSharedPointer in code with
pointer-tracking, you must ensure it gets deleted in code with
pointer-tracking, otherwise the internal safety tracker will be
"leaking" objects. The pointers would never get removed.

And if any new pointer happened to have the same pointer address
(which happens quite often), the tracker code would promptly abort the
application.

With this change, the untracking of the pointer is scheduled by the
same code that creates the tracking. This is done by "abusing" the
custom deleter code:

- for the QSharedPointer that used ExternalRefCountWithDestroyFn
  already, we intercept the call to the destroy function and call the
  untracking function
- for a normal QSharedPointer, we use the "normalDeleter" function as
  custom deleter and chain up above

Note: the autotest only *really* works in release mode. Otherwise
functions don't get inlined and do get merged by the linker.

Reviewed-By: Bradley T. Hughes