Fix deadlock in the fsevents file watcher.
I was making an assumption that one could call CFRunLoopStop and expect
that the runloop would stop very soon afterwards. This is a bit naive
and can result in situations where we end up running past and entering
run again before everything is finished. We now make sure that we don't
leave stop() until the we really have stopped the other thread.

Reviewed-by: mbm