Complete the implementation of QScriptValueIterator with JSCore
This new implementation of QScriptValueIterator passes all the tests.

QScriptValueIterator uses an linked list instead of
JSC::PropertyNameArray so the list can be modified by ::remove() and
to be able to add internal properties for the strings and arrays.

Structure::getPropertyNames() has been modified to not show the
property from the prototype.

Reviewed-by: Kent Hansen