Fixes failed queries when mysql auto-prepares
Queries like "Show Index" etc, fail on mysql when automatically prepared
due to a bug in several versions of mysql. Basically anything but a
select query will fail. This fixes this by making the user explicitly
prepare the query if they want to, and the blame then falls on them
if they try and prepare a statement likely to fail.  This fix also
seems to result in a speedup for single-execution queries, possibly
due to reduction in roundtrip communications.

All autotests pass & behaviour conforms to documentation.

Task-number: 252450, 246125