Animations: better handling of the timer
It could be that if you start lots of animations, they would not be
started at the same time and even have their first tick to late.
So we needed to only transfer the started animation when receiving the
start/stop timer tick.
In addition, if the animation timer is already active we don't restart
it. This would cause no animation to receive ticks if lots of them are
started.