Revert "Remove currentChildBeingDeleted from QObjectPrivate."
This reverts commit 128717b171f01c82e5f0fb83f5923d4f7b9cfc10.

The change broke Alien because QWidget no longer has the native window
handle in its own private. It needs to access the top-level window (or
the one with the native handle) in order to perform some operations.

The problem is that it needs to do that in the destructor. And we
cleared the parent before the destructor was run. Technically
speaking, reverting is the correct thing to do, since the parent
*still* exists when the child is deleted.