Restore the old name for QObjectDeletionNotification.
If we're going to do changes to simplify some of the Declarative UI
enablers, let's do them all at once. That way, the adaptation
necessary happens only once, as opposed to every time we do the
changes.

I tried not to break by leaving a typedef, but obviously the
QObjectPrivate member isn't affected by the typedef...

Reviewed-by: Trust Me