Rewrite bindings inside the compiler
To improve execution performance, binding expressions are rewritten as function
closures inside QmlExpression.  To improve startup performance, where possible,
the expressions are rewritten inside the compiler instead of inside
QmlExpression at runtime.

This also has the sideeffect of removing the StoreBinding instruction, as all
bindings are now "compiled".  The QmlBinding::expression() method for rewritten
bindings will now return the rewritten expression instead of the original (which
is lost), but this API is internal anyway.