Fix a bug when activating HideNameFilterDetails on Windows filedialog.
The HideNameFilterDetails was not taken in account for windows
file dialog. This patch fix that. For Windows Vista and later we use
now the new COMMON_ITEM_DIALOG API (in master) where it add always
the filter extension like *.txt so no point to take that flag in account.

Task-number: None, during testing
Reviewed-by: prasanth