Simplify DirectFB clipping
Instead of storing whether or not our clip is dirty and updating it in
every paint operation we'll just update it whenever it changes. This is
more intuitive and should be faster for the common case as well.

Also make sure to implement clip(region).

Prevent multiple updateClip() calls for a single clip operation by
introducing a flag to check whether we recursed into the different
clip() calls.

Reviewed-by: Donald <qt-info@nokia.com>