Add hints that a qmake generated .pro will need to be edited
When qmake run with the -project option to generate a .pro for the
project it is possible that the generated .pro file will need to be
edited and add variables such as "QT" to specify what QT modules
the project needs. To hints to qmake's help and HTML documentation to
that effect.

Reviewed-by: Marius Storm-Olsen