Fix linking of applications that use QFontDatabase::removeApplicationFont.
Application font support requires the platform to implement the private
registerFont() function, but it is also necessary to implement the public
QFontDatabase::removeApplicationFont() function in the platform fontdatabase.

The former is implemented as a stub, indicating an error to the caller. The
latter was missing as a stub, causing link errors in WebKit.

Reviewed-by: Jason Barron