Make sure QWidget::setStyle() relayouts its children.
Changing a style on a widget should invalidate the widget's layout,
since the layouts spacings and margins may depend on the style.

One optimization could be to check to see if the spacing and margins
have changed due to the style change, but the old style is not available
in changeEvent, so we'll keep it simple for now.

RevBy: paul
Task: 256986