Fix extern WebKit build that uses qmake generators.
WebCore.pro's generators use QMAKE_DEL, which when used with del.exe
works if the file is present or not. In the QMAKE_SH branch QMAKE_DEL
is mapped to rm, which complaints if the file to delete doesn't exist.

Similar to del.exe and unix.conf use rm -f instead of rm for QMAKE_DEL
in the QMAKE_SH branch.

Reviewed-by: Miikka Heikkinen <miikka.heikkinen@digia.com>