Use QFile instead of QTemporaryFile in compilerwarning testcase
QTemporaryFile on Windows doesn't open the file as a sharable, and
doens't close the file when you call .close(). So the testcase
fails on Windows with a Sharing Violation when the compiler tries
to compile the file.

By switching to QFile we can at least close the file before letting
the compiler chew on it, and remove it at the end when the testcase
is done. Open the file with Truncate, in case the testcase fails to
remove the file.

Reviewed-by: trustme