Make sure cursorPositionChanged is emitted when cursor position changes
When selecting the text and setting the cursor position, we need to
emit the cursorPositionChanged signal. This seems to be a copy-paste
error from refactoring QLineEdit, as the function call is there in the
original version of setSelection(). The effect was two test breakages
in tst_QDateTimeEdit, because the section and cursor position of the
QDateTimeEdit would be out of sync.

Reviewed-by: Gunnar