Add ChangeFlags to QGraphicsEffect and pass flags to sourceChanged.
This is way more flexible as it allows for extension in the future.
Instead of having several virtual functions, which in most cases when
re-implemented have to do exactly the same, we now have one virtual
function which takes a ChangedFlags parameter do describe what kind of
changes that occurred to the source.