qdbus: short forms to get/set properties
Allow a property to be given as third parameter instead of a method. If
no other parameter is given, the property value is returned; if a fourth
parameter is given, the property is set to the given value.

This allows the syntax

  qdbus service path property

as a short form of

  qdbus service path org.freedesktop.DBus.Properties.Get property

and the syntax

  qdbus service path property value

as a short form of

  qdbus service path org.freedesktop.DBus.Properties.Set property value

If the third parameter is neither a method nor a property, the
error message for the missing method is returned.

Merge-Request: 1143
Reviewed-By: Thiago Macieira