add configure options for (not) building the QtScript module
-script (default) and -no-script.

This means we can get rid of the SCRIPT feature from qfeatures,
since it's now handled by the new configure variable.

It also allows us to get rid of all the QT_NO_SCRIPT ifdefs from
the source files, since qmake isn't going to include those files
for compilation when you configure with -no-script.

The QtScriptTools module will be disabled if the QtScript module
is not built. You'll have to build the old QtScript back-end
(will be made available in a separate package), then build the
QtScriptTools module yourself.

Reviewed-by: Simon Hausmann