QGraphicsAnchorLayout: Enable graph simplification
This commit enables graph simplification in its simplest way, which is
to simplify the graph, run simplex and then restore it back.

It also adds code to handle cases where the simplification handles
the whole graph, and that therefore no simplex needs to be run.
Additionally, the updateChildrenSizes() method is called on every
root anchor in order to propagate the new sizes down the tree.

A know issue appeared after this commit, as the quote below explains:

=====
Q_ASSERT error in "Parallel" test

That's something that started happening in the Parallel test
(tests/auto/qgraphicsanchorlayout) after my commit of today, that enables
simplification. I haven't been able to investigate it deeply but that's what
I know about it:

- It's triggered when the layout is destroyed, more precisely, from within
"deleteLayoutEdges". What happens is that the layout structural anchors are
missing at that point.

- I suspect there's something to do with the simplifyGraph() and/or
restoreSimplifyGraph() code. I say that because I tried moving the
"restoreSimplifiedGraph" call from the end of the calculateGraph(s) method
to the line exactly below the "simplifyGraph()" and the error still happens.
That means that simplifying and then immediately restoring the graph might
not leave it to the point where it was before the simplification.
====

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>