In a .pro file the include() function does not warn if specified file can not be found change that behavior to warn by default.
Currently the default behavior of include() in a .pro file is not to
warn if the supplied file argument can not be found which can lead to
hard to find build errors. The include() will now,by default, warn if
the specified file can not be found.

If a warning is not required because the included file is optional then
example use in the .pro file:
     include(SomePriFile.pri", "", true)

Task-number:259398
Reviewed-by:Marius Storm-Olsen