Fix default QWidget size on X servers with Xinerama
The default widget size is normally calculated from the root window size, but when running on an X server with Xinerama enabled, the root window spans all screens, making the widget far to large.

Use the default screen's size instead to give the widget a size that
should fit on a single screen.