Make chooseConfig work on some EGL implementations
According to the spec, if we pass a 0 out pointer,
EGL should tell us how many configurations are available.
However, we pass a 0 out pointer, but say that it's 256 elements big,
it confuses some implementations. Fix that by passing a 0 out pointer
and saying that it has space for 0 elements. Now we correctly get the
amount of available configs.

Reviewed-by: Rhys Weatherley <rhys.weatherley@nokia.com>