Fixed KERN-EXEC 3 crash in qdesktopwidget for S60.
QDesktopWidgetPrivate::rects etc variables are shared between
QDesktopWidgetPrivate objects and private objects maintains
refcount to data.

We called cleanup for private object directly from QDesktopWidget,
which caused the shared private data to be freed after some instance
of QDesktopWidget was destroyed. Since we bypassed the refcount also
the data for other QDesktopWidgets was invalidated and this cauyse crash
when resizeEvent was received.

There is no need to call QDesktopWidgetPrivate::cleanup directly since
when QDesktopWidget is deleted, it will call destructor for
QDesktopWidgetPrivate which deletes the shared data based on ref count.