Switched QtNetworkSetting to use 'qt-test-server' config by default.
Qt/Master uses qt-test-server config by default but our S60/master
was configure to use different server by default. The assumption is
that this caused several QtNetwork autotests to fail when testing
QtS60 branch on other configurations in QtSW premises. This because
the old test server was very likely unavailable...