Fix line number and arguments in QScriptContext::toString
For arguments, we always need to skip the implicit 'this' argument

For line number, we walk thought the stack frames from the top to find
the one above the requested one, which contains the returnPC we need.

Also fixed a crash because QScriptContext::parentContext would have
returned a pointer with flags inside.

Reviewed-by: Kent Hansen