Implement advanced pointer handling on S60
Since we only get one pointer event at a time, we need to keep a list of
all known touch points in QApplicationPrivate (otherwise the QTouchEvent
won't contain enough points). The QApplication machinery can handle
having inactive touch-points in the list, so at the moment we don't
clear the list.

We treat PointerNumber zero as the primary touch point, and only send
regular mouse events for that pointer, never for the others.

Reviewed-by: Jason Barron