RVCT didn't like a test bool variable exported directly from QtGui, causing applications to panic with KERN-EXEC 3. Fixed by providing accessor for the bool in question instead of exporting it directly.
Note: Unknown why exactly this panic occurs, as exporting bools works
just fine when tested with small test executable + dll. Something
in QtGui causes this to fail; I suspect it might be the massive
code segment size and relocation of the data segment that is
contributing to this problem.

RevBy: Jason Barron