Wrote an autotest to expose a bug in the GLib event dispatcher.
Zero timers don't fire on the first pass in the GLib event
dispatcher. Ideally I should fix the bug of course, but time doesn't
permit at the moment. Submitting this test instead as a "reminder".

The UNIX event dispatcher passes the test, and it also passes if
moved to the end of the slots.

Task:     259505