Frame property for QDateTimeEdit can be set independently of calendar popup
A frame was always drawn around the QDateTimeEdit editor if a popup
calendar had been set. QStyleOptionsComboBox options are being set in
paintEvent and initialised from the properties of QStyleOptionsSpinBox but
were missing the frame bool property. Now, if the user sets a frame on the
QDateTimeEdit, this property will be consistent with setFrame() property
of the QDateTimeEdit widget.