Ugly round corners when no border is drawn
When specifying round corners with QStyleSheetStyle and no border-width
specified, the round corners were not rendered with antialiasing. Furthermore,
if border-width was set to 0, part of the border was rendered in discordance
with CSS3.

The background in now rendered directly instead of drawing a clipped
rectangle. The actual border width is checked before rendering.

A test has been added at tests/auto/uiloader/baseline/css_borderradius_allwidgets.ui

Task-number: 230362
Reviewed-by: olivier