Fix compilation of unix file engine.
Add a bracket to the right place and re-arrange the code a little.
The indenting was wrong here in the non-Symbian case since the
if (exists) did not have curly bracket. Also copy the union of
the Hidden flag into the Symbian #ifdef to make this code slightly
more readable.