QGraphicsAnchorLayout: fix merging of preferredSizeHint
In simplification, the preferred size of a parallel group should get the
maximum preferred size (and bound it to the maximum size). This follows
the principle that for an layout item: 'growing is better than
shrinking', and is consistent with what our linear program also do.

This patch also make sure that the sizeAt* are properly initialized to
the preferred size, the idea is if there's no restriction upon an
anchor, it'll keep it's preferred size no matter what.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Artur Duque de Souza <artur.souza@openbossa.org>