QGraphicsAnchorLayout: add a few more tests
Adding tests that were helpful during debugging simplification

- example: a replica of the anchorlayout example, which has a few interesting
  cases.
- parallel2: a small pararell test case, subset of example.
- snakeOppositeDirections: snake but with connections in the opposite
  directions (this will end up with anchors with opposite directions).
- fairDistributionOppositeDirections: same as before, but in the end,
  this should be taken care by our API which makes internally the same as
  fairDistribution.

Some work on adding QCOMPARE() statements may still be needed for this
examples.

Note that is interesting to notice (maybe we can make this available
for the test checking automatically) whether the simplex is being used
or not when simplification is active.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>