QGraphicsAnchorLayout: fix expected values for parallel test
We changed the parallel example because the expected values for
preferred size are wrong. Before we had a parallel anchor ("d|e") that
should have the maximum preferred of the both children, so this end up
being 170. This parallel is in a sequence with another item ("b") with
a preferred of 150 ending up with 320.

This group is in parallel with "c" that has a maximum of 300. So in
the preferred size of the layout, the group was constraint to 300,
which would be proportionally distributed between "b" and "d|e", and
50%/50% as the test expected.

The proportional distribution is a feature and we changed the test to
illustrate that more. We gave a bump in "c" maximum to 350 and "b"
maximum to 300, so the preferred is 100% for both "b" and "d|e", and
the proportional feature shows up in the maximum allocation, which is
not 100% for the group (b and d|e).

In the group, the maximum possible is 500 but the layout only allows
350, so the sizes need to be proportionally, but in relation not to
zero, but to the preferred which is 320 (this is the same logic used
for setGeometry()). So the growth is 30 / (500 - 320) = 30 / 180 = 1/6.

This 1/6 should be the factor given to the children (b and d|e),
taking their preferred sizes as starting points.

Note that the test still fails, but serve as a milestone to see when
the feature is correct implemented.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>
Reviewed-by: Artur Duque de Souza <artur.souza@openbossa.org>