Fix the painting of QComboBox on Mac when the rect is not at the origin
The style was assuming that the combo box is painted at (0,0). This is
not the case when the painting is done in the delegate of an item view.

The offset of the rect is now taken into account to paint the style.

HIRect has been replaced by QRect when it make sense.

Task-number: 00026815
Reviewed-by: Richard Moe Gustavsen