Avoid wrapping outside word boundaries in QTextDocument unless necessary
If you have a floating object which affects the width available to the
text, we need to recalculate the width of the text line. In the code,
the setLineWidth() call to do this would by default have WrapAnywhere
as its wrap mode, even when this was not necessary. The code has now
been moved so that WrapAnywhere is only used if we try to set the
line width to match the available width and detect that the text is
too wide (the natural text width exceeds the available space.)

Task-number: 240325
Reviewed-by: Simon Hausmann