Autotest for correct word wrapping on text next to floating object
We set the document's page size to be large enough to contain an
image which is 100 pixels wide and one, but not two, instances of the
word 'Foobar'. We then render HTML which contains a string with repeated
occurrences of the word 'Foobar' next to a floating, right-aligned image
which is 100 pixels wide. The layout should break on word boundaries,
since this is the default in QTextDocument, and thus each text line
should contain one instance of the word 'Foobar'.

Task-number: 240325