Improve memory usage of the connectionlists inside QObject
... by not allocating space for slots in the vector.

Before, the vector uses the signal index as index. The problem is that
the slots and signal are mixed in the same index space.

We solve the problem by having a different index space for the signal in
the connectionlists vector.  All we need to do is to add the information
about the number of signals in the moc.

Also, we are not connecting to cloned signal but only to the orginial
ones.  For example,  destroyed(QObject * = 0) would generate two signal,
we now only connect to the first one.

This also improve a little bit the performence while activating signals
since it removed one call to indexOfMethod.

Reviewed-by: Brad