Use LIBS_PRIVATE on Mac and X11.
On the Mac, it means "-framework ApplicationServices -framework Carbon
-framework AppKit" are no longer part of the default LIBS in Qt
applications. This required a lot of fixes where we used Mac-specific
code in Qt.

On X11, it was very straightforward, because we apparently use very
little of X11 outside QtGui.

I haven't changed the Windows-specific LIBS paths, because I don't
know how Windows behaves. Windows has DLLs, but it links to static
"import" libraries. So is it static linking or dynamic linking?

Reviewed-By: Marius Storm-Olsen