Add support for the orientation of tiff images
The orientation tag was ignored for tiff images. The tag is now used
to rotate the image before providing it to the user.

The orientation of indexed an mono images is done completely manually.

The orientation of other type is done in conjunction to the mirroring
already performed by libtiff.

Task-number: 258526
Reviewed-by: Samuel Rødal