QGraphicsAnchorLayout: calculate vertex positions with simplified graph
When traversing the graph looking for calculating the distances, also
enter inside the complex edges to find (and calculate) the distances for
the vertices that were simplified.

This require a little tweak in the traversal, now we can't always skip an edge
when both vertices are already visited, because complex anchors can hide
vertices inside (a sequential anchor in practice remove vertices from the
full graph).

As a bonus we now pass on the 'example' test (as expected) and the
example now works fine.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@openbossa.org>