QGraphicsAnchorLayout: fix calculation of sizeAt* values for Sequential
This commit implements what's described in the previous commit

    QGraphicsAnchorLayout: fix expected values for parallel test

When filling the sizeAt* values (the three points used for interpolation
when setting geometry), now sequential anchors distribute the sizes
to the children in a fair way, i.e. proportionally in relation to the
existing min/pref/max hints.

Each value is defined in relation of the pref (either as a shrinkage in
pref or a grow in pref). In both cases the shrinking/growing factor is
the same for all children.

When we implement support for QSizePolicies, this distribution might be
changed by setting size policies for a certain item or anchor.

This makes two tests work -- so no more expected fail. Also fixed a typo
in one test's expected results.