Fixes to Symbian native engine for QFileSystemWatcher.
The wrong version of cancel was used, resulting in all filesystem
notifications being cancelled when one path was removed from the engine.
For directory notifications, changed ENotifyAll to ENotifyEntry to
reduce the spurious directory changed notifications which come from
file modifications within that directory.
On Symbian OS, opening a file for overwrite (truncate mode in Qt terms)
will cause a directory change notification, the OS considers it to be
the same type of event as creating a new file for the first time.
Other platforms presumably do not have this behaviour otherwise the auto
test would fail there as well.

Reviewed-by: miikka heikkinen