Fix autotest to accept Symbian OS behaviour as a pass.
Symbian generates a directory change event when a file is overwritten in
that directory.
Autotest was asserting that no change event is generated in this
situation (documentation allows it in the signal description)
So the test now allows 0 or 1 dir changed events, to allow existing
platforms and symbian to both pass.

Reviewed-by: brad